thread: make thread_context::s_main portable
On some platforms (e.g. IBM Z) the ABI requires that arguments are
sign/zero extended to register width as appropriate for their type.
This means that signed and unsigned arguments cannot be used
interchangeably since the compiler can assume that it doesn't have
to explicitly zero extend an unsigned argument or sign extend a
signed argument.

s_main is called indirectly and the caller assumes that it takes
signed ints, so s_main must accept signed int arguments. If it
expects unsigned ints then the thread_context pointer will not be
rematerialized correctly when lo&(1<<31) != 0.

Message-Id: <20170905234423.82554-1-mike.munday@ibm.com>

+++ core/thread.cc
@@ -277,8 +277,8 @@ thread_context::reschedule() {
 }
 
 void
-thread_context::s_main(unsigned int lo, unsigned int hi) {
-    uintptr_t q = lo | (uint64_t(hi) << 32);
+thread_context::s_main(int lo, int hi) {
+    uintptr_t q = uint64_t(uint32_t(lo)) | uint64_t(hi) << 32;
     reinterpret_cast<thread_context*>(q)->main();
 }
